employee done

+++ src/screens/Employee/index.js
@@ -113,22 +113,24 @@ const EmployeeScreen = () => {
 							flexDirection: 'column',
 						}}
 						onClick={() => {
-							var axios = require('axios');
-							var config = {
-								method: 'delete',
-								url: `${host.host}/profile`,
-								headers: {
-									Credentials: `Bearer ${window.localStorage.getItem(
-										'dbisToken'
-									)}`,
-								},
-							};
-							axios(config)
-								.then((res) => {
-									console.log(res.data);
-									window.location = '/signup';
-								})
-								.catch((err) => console.log(err));
+							if (window.confirm('Are you sure to delete your profile')) {
+								var axios = require('axios');
+								var config = {
+									method: 'delete',
+									url: `${host.host}/profile`,
+									headers: {
+										Credentials: `Bearer ${window.localStorage.getItem(
+											'dbisToken'
+										)}`,
+									},
+								};
+								axios(config)
+									.then((res) => {
+										console.log(res.data);
+										window.location = '/signup';
+									})
+									.catch((err) => console.log(err));
+							}
 						}}
 					>
 						<h2 style={{ color: '#751919', textDecoration: 'underline' }}>
